requirements.txt
Django                        
django-allauth                
djangorestframework          
djangorestframework-simplejwt
cryptography
drf-yasg
PyJWT
asgiref
mysqlclient
pytz
python-dateutil
sqlparse
PyJWT
requests
oauthlib
requests-oauthlib
gunicorn==20.1.0
openai